% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-data.R
\name{get_netflix_weekly_data}
\alias{get_netflix_weekly_data}
\title{Scrape Netflix Top 10 Data for a Specific Week}
\usage{
get_netflix_weekly_data(week_date)
}
\arguments{
\item{week_date}{String or Date. The Sunday ending the week (Format: "YYYY-MM-DD").}
}
\value{
A dataframe containing the Top 10 for all 4 categories.
}
\description{
Scrape Netflix Top 10 Data for a Specific Week
}
