% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{import_netflix_excel}
\alias{import_netflix_excel}
\title{Import Official Netflix Excel Dataset}
\usage{
import_netflix_excel(file_path)
}
\arguments{
\item{file_path}{String. Local path to 'all-weeks-global.xlsx' OR a direct URL.}
}
\value{
A cleaned dataframe ready for analysis.
}
\description{
Import Official Netflix Excel Dataset
}
