% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.R
\name{ler_netflix}
\alias{ler_netflix}
\title{Ler tabela com dados de histórico do netflix}
\usage{
ler_netflix(path)
}
\arguments{
\item{path}{Caminho para arquivo com histórico}
}
\value{
Uma \code{\link{tibble}} com as colunas:

data: data do registro

serie: Série ou Filme

titulo: Tïtulo da série ou do filme

temporada: Temporada da série ou NA para filme

episodio: Episório da série ou NA para filme
}
\description{
Ler tabela com dados de histórico do netflix
}
\examples{
# ler_netflix("meu_historico.csv")
}
