% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-data.R
\name{scrape_multiple_weeks}
\alias{scrape_multiple_weeks}
\title{Scrape Multiple Weeks of Data}
\usage{
scrape_multiple_weeks(start_date, end_date)

scrape_multiple_weeks(start_date, end_date)
}
\arguments{
\item{start_date}{String. Start date (YYYY-MM-DD).}

\item{end_date}{String. End date (YYYY-MM-DD).}
}
\value{
A combined dataframe of all weeks.

A combined dataframe of all weeks.
}
\description{
Scrape Multiple Weeks of Data

Scrape Multiple Weeks of Data
}
